// src/model/api/soundcloud/SoundCloudArtistSearchResponse.ts
export interface SoundCloudArtistSearchResponse {
  avatar_url: string;
  id: number;
  kind: string;
  permalink_url: string;
  username: string;
  first_name: string;
  last_name: string;
  full_name: string;
  city: string;
  description: string;
  country: string;
  track_count: number
  public_favorites_count: number;
  reposts_count: number;
  followers_count: number;
  followings_count: number;
  plan: string;
}